+++ src/pages/keep-safe/index.js
@@ -2,19 +2,19 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import device from 'themes/device'
-import SEO from '../components/containers/seo'
-import Layout from '../components/layout/layout'
-import Image from '../components/elements/image'
-import Hero from '../components/elements/hero.js'
-import Button from '../components/form/button'
-import Modal, { useModal } from '../components/elements/modal'
-import SignupModal from '../components/elements/signup-modal'
-import Container from '../components/containers/container'
-import { Card } from '../components/elements/card.js'
-import { Divider } from '../components/elements/divider'
-import { StyledLink } from '../components/elements/link'
-import { Header, Text } from '../components/elements/typography.js'
-import { localize, WithIntl } from '../components/localization'
+import SEO from '../../components/containers/seo'
+import Layout from '../../components/layout/layout'
+import Image from '../../components/elements/image'
+import Hero from '../../components/elements/hero.js'
+import Button from '../../components/form/button'
+import Modal, { useModal } from '../../components/elements/modal'
+import SignupModal from '../../components/elements/signup-modal'
+import Container from '../../components/containers/container'
+import { Card } from '../../components/elements/card.js'
+import { Divider } from '../../components/elements/divider'
+import { StyledLink } from '../../components/elements/link'
+import { Header, Text } from '../../components/elements/typography.js'
+import { localize, WithIntl } from '../../components/localization'
 
 import Two from 'images/svg/two-factor.svg'
 import DifferentPassword from 'images/svg/different-password.svg'
