+++ src/pages/regulatory/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-import SEO from '../components/containers/seo'
-import Layout from '../components/layout/layout'
-import { localize, WithIntl } from '../components/localization'
+import SEO from '../../components/containers/seo'
+import Layout from '../../components/layout/layout'
+import { localize, WithIntl } from '../../components/localization'
 import Hero from 'components/elements/hero'
 import Container from 'components/containers/container'
 import { Header, Text } from 'components/elements/typography'
